cli: Implement the commands line interface
This is a nice first pass. This does what is minimally
required to work with the ssh auth dir repo.

In order to do so, added some significant backing libraries.

Some unit tests but was mostly manually tested. This is
something to be improved on.

Deauthorization interface / behavior might have to be
improved as a bit unintuitive.

No custom autocompletions yet. This would be a really nice thing
to add.

Git subcommands still unimplemented.